requirements.txt
python-docx
transformers
PyMuPDF
Pillow
accelerate
tiktoken
verovio